+++ FourSix Coffee Timer/Model/Calculator.swift
@@ -11,9 +11,9 @@ import Foundation
 struct Calculator {
     
     var recipe: Recipe?
-    var waterPours = [Double]()
+    var waterPours = [Float]()
     
-    mutating func calculate(_ balance: Balance, _ strength: Strength, with coffee: Double, _ water: Double) {
+    mutating func calculate(_ balance: Balance, _ strength: Strength, with coffee: Float, _ water: Float) {
         let water40 = water * 0.4
         let firstPour = water40 * balance.rawValue
         let secondPour = water40 - firstPour
@@ -23,18 +23,18 @@ struct Calculator {
     
         let water60 = water * 0.6
         let water60Count = strength.rawValue
-        let water60Pour = water60 / Double(water60Count)
+        let water60Pour = water60 / Float(water60Count)
         
         waterPours.append(contentsOf: repeatElement(water60Pour.rounded(), count: water60Count))
         
         recipe = Recipe(waterTotal: water, coffee: coffee, waterPours: waterPours, balance: balance, strength: strength)
     }
     
-    func getPours() -> [Double] {
+    func getPours() -> [Float] {
         return recipe?.waterPours ?? [60, 60, 60, 60, 60]
     }
     
-    func getTotalWater() -> Double {
+    func getTotalWater() -> Float {
         return recipe?.waterTotal ?? 300
     }
